Re arrange files
