+++ src/App.js
@@ -1,6 +1,6 @@
 import { Routes, Route } from 'react-router';
 import Header from './components/Header';
-import HomePage from './pages/HomePage';
+import Home from './pages/Home';
 import Footer from './components/Footer';
 import './App.css';
 
@@ -12,7 +12,7 @@ function App() {
         <Routes>
           <Route
             path="/"
-            element={(<HomePage />)}
+            element={(<Home />)}
           />
         </Routes>
       </div>
