+++ src/pages/About.js
@@ -1,0 +1,13 @@
+import AboutMe from "../components/About/AboutMe";
+
+const About = () => {
+  return (
+  <div>
+    {/* about-me-section */}
+    <AboutMe />
+    
+  </div>
+  )
+}
+
+export default About;
